Improved the user menu and added a settings page
