+++ packages/user/package.json
@@ -47,6 +47,7 @@
     "@lumino/commands": "^1.12.0",
     "@lumino/coreutils": "^1.5.3",
     "@lumino/signaling": "^1.4.3",
+    "@lumino/virtualdom": "^1.8.0",
     "@lumino/widgets": "^1.26.0",
     "react": "^17.0.1"
   },
