+++ packages/user/style/index.css
@@ -9,3 +9,4 @@
 @import url('~@jupyterlab/apputils/style/index.css');
 
 @import url('./base.css');
+@import url('./settings.css');
