+++ packages/user/style/index.js
@@ -9,3 +9,4 @@ import '@jupyterlab/ui-components/style/index.js';
 import '@jupyterlab/apputils/style/index.js';
 
 import './base.css';
+import './settings.css';
